test: move event suites out of src coverage
