+++ tests/events/src-runtime-internal-error.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from "vitest";
-import { RuntimeEventBus } from "../runtime-events.js";
+import { RuntimeEventBus } from "../../src/events/runtime-events.js";
 
 describe("RuntimeEventBus runtime.internal.error", () => {
   it("emits runtime.internal.error when a listener throws", () => {
